Introduce unified simulation command service API
